# swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Wafelbak API",
      "description": "The official documentation of the Wafelbak API",
      "version": "0.1"
    },
    "servers": [
      {
        "url": "https://wafelbak-backend.herokuapp.com"
      }
    ],
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "tags": [
      {
        "name": "Users"
      }
    ],
    "paths": {
      "/API/users/register": {
        "post": {
          "tags": ["Users"],
          "description": "Register new user",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/register"
                },
                "examples": {
                  "0": {
                    "value": "{\r\n\t\"firstName\": \"NewClient\",\r\n\t\"lastName\": \"ToDelete\",\r\n\t\"email\": \"client@gmail.com\",\r\n\t\"password\": \"test00##\",\r\n\t\"birthday\": \"2019-12-27\",\r\n\t\"street\": \"Vroonstallestraat\",\r\n\t\"streetNumber\": 8,\r\n\t\"postalCode\": 9032,\r\n\t\"city\": \"Wondelgem\"\r\n}"
                  },
                  "1": {
                    "value": "{\r\n\t\"firstName\": \"Victor\",\r\n\t\"lastName\": \"Van hulle\",\r\n\t\"email\": \"vanhulle.victor@gmail.com\",\r\n\t\"password\": \"123Victor!\",\r\n\t\"birthday\": \"1997-02-08\",\r\n\t\"street\": \"Woestijnegoedlaan\",\r\n\t\"streetNumber\": 38,\r\n\t\"postalCode\": 9032,\r\n\t\"city\": \"Wondelgem\",\r\n\t\"admin\": true\r\n}"
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "New user successfully registered",
              "content": {
                "application/json; charset=utf-8": {
                  "schema": {
                    "type": "string"
                  },
                  "examples": {}
                }
              }
            },
            "500": {
              "description": "Wrong or missing argument",
              "content": {
                "application/json; charset=utf-8": {
                  "schema": {
                    "type": "string"
                  },
                  "examples": {}
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://wafelbak-backend.herokuapp.com"
            }
          ]
        },
        "servers": [
          {
            "url": "https://wafelbak-backend.herokuapp.com"
          }
        ]
      },
      "/API/users/login": {
        "post": {
          "tags": ["Users"],
          "description": "Login as user or mentor",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/login"
                },
                "examples": {
                  "0": {
                    "value": "{\r\n\t\"email\": \"client@gmail.com\",\r\n\t\"password\": \"test00##\"\r\n}"
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successfully logged in",
              "content": {
                "application/json; charset=utf-8": {
                  "schema": {
                    "type": "string"
                  },
                  "examples": {}
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://wafelbak-backend.herokuapp.com"
            }
          ]
        },
        "servers": [
          {
            "url": "https://wafelbak-backend.herokuapp.com"
          }
        ]
      },
      "/API/users": {
        "get": {
          "tags": ["Users"],
          "description": "Get all users",
          "responses": {
            "200": {
              "description": "Return list of all users"
            }
          },
          "servers": [
            {
              "url": "https://wafelbak-backend.herokuapp.com"
            }
          ]
        },
        "servers": [
          {
            "url": "https://wafelbak-backend.herokuapp.com"
          }
        ]
      },
      "/API/users/client@gmail.com": {
        "get": {
          "tags": ["Users"],
          "description": "Get user by email",
          "responses": {
            "200": {
              "description": "Successfully returned user with corresponding email",
              "content": {
                "application/json; charset=utf-8": {
                  "schema": {
                    "type": "string"
                  },
                  "examples": {}
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://wafelbak-backend.herokuapp.com"
            }
          ]
        },
        "servers": [
          {
            "url": "https://wafelbak-backend.herokuapp.com"
          }
        ]
      },
      "/API/users/id/11": {
        "get": {
          "tags": ["Users"],
          "description": "Get user by ID",
          "responses": {
            "200": {
              "description": "Successfully returned user with corresponding ID",
              "content": {
                "application/json; charset=utf-8": {
                  "schema": {
                    "type": "string"
                  },
                  "examples": {}
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://wafelbak-backend.herokuapp.com"
            }
          ]
        },
        "servers": [
          {
            "url": "https://wafelbak-backend.herokuapp.com"
          }
        ]
      }
    },
    "components": {
      "schemas": {
        "login": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          }
        },
        "register": {
          "type": "object",
          "properties": {
            "firstName": {
              "type": "string"
            },
            "lastName": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "birthday": {
              "type": "string"
            },
            "street": {
              "type": "string"
            },
            "streetNumber": {
              "type": "integer"
            },
            "streetExtra": {
              "type": "string"
            },
            "postalCode": {
              "type": "string"
            },
            "city": {
              "type": "string"
            },
            "admin": {
              "type": "boolean"
            }
          }
        }
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }
  